% Generated by roxygen2 (4.0.1): do not edit by hand
\name{diff.exp}
\alias{diff.exp}
\title{Compute the differential expression for a GSTTDataSet object}
\usage{
diff.exp(object, contrast, name, ...)
}
\arguments{
\item{object}{an object of class GSTTDataSet}

\item{contrast}{the contrast to compare the effects or levels on the expression, see \code{DESeq2::results}}

\item{name}{the name of the main effect, if not using contrast}
}
\value{
an object of class GSTTDataSet with the results and col.groups slots filled
}
\description{
\code{diff.exp} runs DESeq for the dataset
}
\seealso{
\code{DESeq2::results}
}

