% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.gstt.results}
\alias{get.gstt.results}
\title{Compute the p-values for the enrichment of differentially expressed genes in the given gene sets/}
\usage{
get.gstt.results(object, gene_sets)
}
\arguments{
\item{object}{of class GSTTDataSet}

\item{gene_sets}{a list of gene sets}
}
\value{
object of class GSTTResults
}
\description{
Compute the p-values for the enrichment of differentially expressed genes in the given gene sets/
}

