% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getContrast}
\alias{getContrast}
\title{Get the numerical contrast vector for the dataset}
\usage{
getContrast(object, contrast, name)
}
\arguments{
\item{gstt}{an object of class DESeqDataSet}
}
\description{
Get the numerical contrast vector for the dataset
}
\examples{
library(DESeq2)
example(DESeq)
getContrast(dds)
}
\seealso{
\code{DESeq2:::cleanContrast}
}

