% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getGroupsFromCharContrast}
\alias{getGroupsFromCharContrast}
\title{Get the list of two groups of samples for the contrast.
The groups of samples can then be used to get the balanced permutations.}
\usage{
getGroupsFromCharContrast(object, contrast, expanded)
}
\arguments{
\item{object}{of class DESeqDataset}

\item{contrast}{a vector c('contrastFactor','Level1','Level2') if we want do Level1 vs Level2}
}
\description{
Get the list of two groups of samples for the contrast.
The groups of samples can then be used to get the balanced permutations.
}

