% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GSTTDataSetFromMatrix}
\alias{GSTTDataSetFromMatrix}
\title{get a GSTTDataSet from a matrix of transcript abundance counts}
\usage{
GSTTDataSetFromMatrix(countData, colData, design, ...)
}
\arguments{
\item{countData}{a matrix of transcript abundance counts}

\item{colData}{a data frame where rows represents samples (columns of contData) and columns represent properties of samples}
}
\value{
a GSTTDataSet object
}
\description{
get a GSTTDataSet from a matrix of transcript abundance counts
}

