% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qqpermuted}
\alias{qqpermuted}
\title{plot the q-q plot of the pvals of the permuted results}
\usage{
qqpermuted(object, pindex = "all", plot.it = T)
}
\arguments{
\item{object}{of class GSTTDataSet with the \code{permuted.results} slot filled}
}
\description{
plot the q-q plot of the pvals of the permuted results
}

