% Generated by roxygen2 (4.0.1): do not edit by hand
\name{diff.exp}
\alias{diff.exp}
\title{Compute the differential expression for all permutations for a GSTTDataSet object}
\usage{
diff.exp(object)
}
\arguments{
\item{object}{an object of class GSTTDataSet}
}
\value{
an object of class GSTTDataSet with the p-values computed for all sample permutations
}
\description{
\code{diff.exp} runs DESeq for the dataset and all its sample permutations needed to get
the gene set significance levels.
}

